// graphql/schema/generated/fragments/notebook.fragment.ts
/* Do not edit this file. It was generated programmatically. */

import { Notebook } from "@/graphql/schema/generated/models/notebook.model";
import { gql } from "@apollo/client";

export const notebookFragment = gql`
  fragment NotebookFragment on Notebook {
    __typename
    id
    userId
    title
    description
    public
    createdAt
    updatedAt
    archivedAt
  }
`;

export type NotebookFragment = Pick<
  Notebook,
  | "__typename"
  | "id"
  | "createdAt"
  | "updatedAt"
  | "archivedAt"
  | "userId"
  | "title"
  | "description"
  | "public"
>;