// graphql/schema/generated/fragments/note.fragment.ts
/* Do not edit this file. It was generated programmatically. */

import { Note } from "@/graphql/schema/generated/models/note.model";
import { gql } from "@apollo/client";

export const noteFragment = gql`
  fragment NoteFragment on Note {
    __typename
    id
    userId
    notebookId
    title
    body
    public
    createdAt
    updatedAt
    archivedAt
  }
`;

export type NoteFragment = Pick<
  Note,
  | "__typename"
  | "id"
  | "createdAt"
  | "updatedAt"
  | "archivedAt"
  | "userId"
  | "notebookId"
  | "title"
  | "body"
  | "public"
>;